Move context aware keyboard shortcut into its own file.
